fixes #1
there was a problem with the indexing of two vectors whose results are
used to split apart the combined columns in the raw UN data. th issue
was fixed by separating out the indexing for one of the vectors, and
manually iterating over it.

another downstream issue occurred with the naming of columns after the
expansion of the columns, this appears to be working as well, though it
has only been informally tested.